Function to load editor content to simulator
The editor.ui has been changed and recompiled to ui_editor.py.

This change also allows erroneous lines to be highlighted when loaded
directly from the editor to the simulation, that's why DCError has a new
field to record the line number (felt better than parsing the error
message for information that we should have available).

The red background in the flawed line will disappear once you move the
cursor.