Fix editor colors to black on white
On some themes a style might have white text on black background. Since
we use custom colors for syntax highlight and current line highlight,
text was hard to read when the system theme used white text. To fix this
it's easier to always have the same colors instead of defining different
themes for the editor.