Rewrite assemble method
The old code was a huge mess, this one is slightly better. It first
tokenizes the input, then iterates over each token and defines the
labels. In the last round it iterates over the instructions again and
replaces the label names with the right line numbers.

IDict is now obsolete and not needed anymore, it was removed from
dc.util (it wasn't really good either).